+++ setup/dist-collect-files.sh
@@ -18,7 +18,7 @@ rm    "${DIR}"/../../neucore/backend/var/logs/.gitkeep
 mv    "${DIR}"/../doc "${DIR}"/../../neucore/doc
 rm -r "${DIR}"/../../neucore/doc/screenshots
 rm    "${DIR}"/../../neucore/doc/api-examples/.gitignore
-rm    "${DIR}"/../../neucore/doc/API.md.tpl
+rm    "${DIR}"/../../neucore/doc/API.tpl.md
 rm    "${DIR}"/../../neucore/doc/er-model.mwb
 
 mkdir "${DIR}"/../../neucore/setup